+++ src/main/java/spring_boot_security/dao/RoleDao.java
@@ -9,5 +9,5 @@ public interface RoleDao {
 
     List<Role> getAllRoles();
 
-    Set<Role> getRoleById(List<Integer> rolesId);
+    Set<Role> getRoleById(List<Long> rolesId);
 }
