+++ src/main/java/spring_boot_security/service/UserService.java
@@ -13,11 +13,11 @@ public interface UserService extends UserDetailsService {
 
     void saveUser(User user, Set<Role> roles);
 
-    void updateUser(User user, int id, Set<Role> roles);
+    void updateUser(User user, long id, Set<Role> roles);
 
     void updateUser(User user);
 
-    void deleteUser(int id);
+    void deleteUser(long id);
 
-    User getUserById(int id);
+    User getUserById(long id);
 }
